+++ src/mastra/tools/statistics/helpers.ts
@@ -93,7 +93,7 @@ export async function getDueCardsStats(
       COALESCE(SUM(rs.repetitions), 0) as total_reviews
     FROM cards c
     LEFT JOIN review_state rs ON c.id = rs.card_id
-    WHERE c.owner_id = $1 AND c.active = true
+    WHERE c.owner_id = $1::bigint AND c.active = true
   `;
 
   const result = await pool.query(statsQuery, [owner_id, today, tomorrow]);
@@ -146,7 +146,7 @@ export async function getRetentionStats(
         COUNT(CASE WHEN rl.reviewed_at >= $4 AND rl.grade >= $2 THEN 1 END) as successful_last_30_days
       FROM review_log rl
       JOIN cards c ON rl.card_id = c.id
-      WHERE c.owner_id = $1 AND c.active = true
+      WHERE c.owner_id = $1::bigint AND c.active = true
     ),
     card_maturity AS (
       SELECT 
@@ -154,7 +154,7 @@ export async function getRetentionStats(
         COUNT(CASE WHEN rs.interval_days < 21 AND rs.interval_days > 0 THEN 1 END) as young_cards
       FROM review_state rs
       JOIN cards c ON rs.card_id = c.id
-      WHERE c.owner_id = $1 AND c.active = true
+      WHERE c.owner_id = $1::bigint AND c.active = true
     )
     SELECT 
       rs.*,
@@ -227,7 +227,7 @@ export async function getStreakStats(
       COUNT(*) as reviews_count
     FROM review_log rl
     JOIN cards c ON rl.card_id = c.id
-    WHERE c.owner_id = $1 AND c.active = true
+    WHERE c.owner_id = $1::bigint AND c.active = true
     GROUP BY DATE(rl.reviewed_at)
     ORDER BY review_date DESC
   `;
@@ -245,10 +245,10 @@ export async function getStreakStats(
       SELECT DATE(reviewed_at) as date, COUNT(*) as daily_reviews
       FROM review_log rl2
       JOIN cards c2 ON rl2.card_id = c2.id
-      WHERE c2.owner_id = $1 AND c2.active = true
+      WHERE c2.owner_id = $1::bigint AND c2.active = true
       GROUP BY DATE(reviewed_at)
     ) daily_counts ON DATE(rl.reviewed_at) = daily_counts.date
-    WHERE c.owner_id = $1 AND c.active = true
+    WHERE c.owner_id = $1::bigint AND c.active = true
   `;
 
   const [reviewDatesResult, overallStatsResult] = await Promise.all([
@@ -367,7 +367,7 @@ export async function getEaseHistogram(
     SELECT rs.ease_factor
     FROM review_state rs
     JOIN cards c ON rs.card_id = c.id
-    WHERE c.owner_id = $1 AND c.active = true
+    WHERE c.owner_id = $1::bigint AND c.active = true
     ORDER BY rs.ease_factor
   `;
