+++ src/mastra/tools/statisticsTools.ts
@@ -265,7 +265,7 @@ export const getStreakStatsTool = createTool({
           COUNT(*) as reviews_count
         FROM review_log rl
         JOIN cards c ON rl.card_id = c.id
-        WHERE c.owner_id = $1 AND c.active = true
+        WHERE c.owner_id = $1::bigint AND c.active = true
         GROUP BY DATE(rl.reviewed_at)
         ORDER BY review_date DESC
       `;
@@ -283,10 +283,10 @@ export const getStreakStatsTool = createTool({
           SELECT DATE(reviewed_at) as date, COUNT(*) as daily_reviews
           FROM review_log rl2
           JOIN cards c2 ON rl2.card_id = c2.id
-          WHERE c2.owner_id = $1 AND c2.active = true
+          WHERE c2.owner_id = $1::bigint AND c2.active = true
           GROUP BY DATE(reviewed_at)
         ) daily_counts ON DATE(rl.reviewed_at) = daily_counts.date
-        WHERE c.owner_id = $1 AND c.active = true
+        WHERE c.owner_id = $1::bigint AND c.active = true
       `;
 
       logger?.info('📝 [GetStreakStats] Executing streak statistics queries');
@@ -447,7 +447,7 @@ export const getEaseHistogramTool = createTool({
         SELECT rs.ease_factor
         FROM review_state rs
         JOIN cards c ON rs.card_id = c.id
-        WHERE c.owner_id = $1 AND c.active = true
+        WHERE c.owner_id = $1::bigint AND c.active = true
         ORDER BY rs.ease_factor
       `;
